+++ ChessApp/ChessApp/Main.swift
@@ -7,14 +7,19 @@
 
 import Foundation
 
+typealias Board = [[PawnType]]
+
 func main() {
-    
     // MARK: - 체스 시작 초기화
     // "체스 보드를 초기화했습니다." 출력
     // 8 x 8 board 초기화 (1,2-rank는 흑색 체스말이, 7,8-rank는 백색 체스말이 위치)
     
     let boardManager = BoardManager(size: 8)
-    // 체스 보드 출력
+    let printManager = PrintManager(board: boardManager.board)
+    
+    // 초기 체스 보드 출력
+    printManager.printBoard()
+    
     // "명령을 입력하세요>" 출력
     
     
@@ -24,14 +29,12 @@ func main() {
     
     
     // MARK: - [처리/계산] board 저장값 수정
-    boardManager.printBoard()
     boardManager.updateBoard(currentPosition: Position(rank: "A", file: 2), updatePosition: Position(rank: "A", file: 3))
-//    board[A][3] = board[A][2]
-    boardManager.printBoard()
     
     
     // MARK: - [형식, 출력] 수정된 board 출력
     // 상대편 말을 잡는 경우 수정된 board 출력
+    printManager.printBoard()
 }
 
 main()
